Moved case service types to types directory since they were building
