+++ api/src/middlewares/services/CaseService/types/FormMetadata.ts
@@ -1,0 +1,16 @@
+import { CertifyCopyFormType, FormType, MarriageFormType, PayMetadata } from "../../../../types/FormDataBody";
+
+export interface FormMetadata {
+  reference: string;
+  payment?: PayMetadata;
+  type: FormType;
+}
+
+export interface MarriageFormMetadata extends FormMetadata {
+  type: MarriageFormType;
+  postal?: boolean;
+}
+
+export interface CertifyCopyFormMetadata extends FormMetadata {
+  type: CertifyCopyFormType;
+}
