+++ api/src/middlewares/services/CaseService/types/PaymentViewModel.ts
@@ -1,0 +1,10 @@
+export type PaymentViewModel = {
+  id: string;
+  status: string;
+  total: string;
+  url: string;
+  allTransactionsByCountry: {
+    url: string;
+    country: string;
+  };
+};
