+++ api/src/middlewares/services/CaseService/types/ProcessQueueData.ts
@@ -1,0 +1,17 @@
+import { FormField } from "../../../../types/FormField";
+import { CertifyCopyFormMetadata, FormMetadata, MarriageFormMetadata } from "./FormMetadata";
+
+interface ProcessQueueDataBase {
+  fields: FormField[];
+  metadata: FormMetadata;
+}
+
+export interface CertifyCopyProcessQueueData extends ProcessQueueDataBase {
+  metadata: CertifyCopyFormMetadata;
+}
+
+export interface MarriageProcessQueueData extends ProcessQueueDataBase {
+  metadata: MarriageFormMetadata;
+}
+
+export type ProcessQueueData = CertifyCopyProcessQueueData | MarriageProcessQueueData;
